requirements.txt
django==5.1.15
django-ninja==1.3.0
psycopg2-binary==2.9.10
bcrypt==4.2.1
PyJWT==2.12.0
python-dotenv==1.0.1
pydantic[email]==2.10.4
bandit==1.7.10
pip-audit==2.7.3
